+++ superflore/utils.py
@@ -36,6 +36,46 @@ else:
         return response.read()
 
 
+# TODO(allenh1): This is a blacklist of things that
+# do not yet support Python 3. This will be updated
+# on an as-needed basis until a better solution is
+# found (CI?).
+
+no_python3 = ['tf']
+
+
+def warn(string):
+    print(colored('>>>> {0}'.format(string), 'yellow'))
+
+
+def ok(string):
+    print(colored('>>>> {0}'.format(string), 'green'))
+
+
+def err(string):
+    print(colored('!!!! {0}'.format(string), 'red'))
+
+
+def info(string):
+    print(colored('>>> {0}'.format(string), 'cyan'))
+
+
+def make_dir(dirname):
+    try:
+        os.makedirs(dirname)
+    except:
+        pass
+
+
+def get_pkg_version(distro, pkg_name):
+    pkg = distro.release_packages[pkg_name]
+    repo = distro.repositories[pkg.repository_name].release_repository
+    maj_min_patch, deb_inc = repo.version.split('-')
+    if deb_inc != '0':
+        return '{0}-r{1}'.format(maj_min_patch, deb_inc)
+    return maj_min_patch
+
+
 def download_yamls():
     global base_yml
     global python_yml
@@ -46,11 +86,11 @@ def download_yamls():
     python_yaml = "{0}/python.yaml".format(base_url)
     ruby_yaml = "{0}/ruby.yaml".format(base_url)
 
-    print(colored("Downloading latest base yml...", 'cyan'))
+    info("Downloading latest base yml...")
     base_yml = yaml.load(get_http(base_yaml))
-    print(colored("Downloading latest python yml...", 'cyan'))
+    info("Downloading latest python yml...", 'cyan')
     python_yml = yaml.load(get_http(python_yaml))
-    print(colored("Downloading latest ruby yml...", 'cyan'))
+    info("Downloading latest ruby yml...", 'cyan')
     ruby_yml = yaml.load(get_http(ruby_yaml))
